LayoutTests/media/media-controls-accessibility-expected.txt


This tests that media controls have correct accessibility information.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
elapsedTimer.description: AXDescription: Elapsed 00:00
remainingTimer.description: AXDescription: Remaining -00:06

muteButton.description: AXDescription: Mute
muteButton.role: AXRole: AXCheckBox
muteButton.stringValue: AXValue: 0
press muteButton
muteButton.stringValue: AXValue: 1

timeline.value: 0.5
timeline.value: 0.0

RUN(video.load())
EVENT(canplaythrough)
EVENT(seeked)
EVENT(seeked)
END OF TEST
